Move fitting into 'check' and use in 'validate'.
Make checkAstrometry and checkPhotometry run fits
of expected error behavior from astrometric and photometric.
