+++ python/lsst/validate/drp/validate.py
@@ -528,16 +528,20 @@ def runOneFilter(repo, visitDataIds, brightSnr=100,
     mmagerr = 1000*magerr
     mmagrms = 1000*magrms
 
-    checkAstrometry(struct.snr, dist, match,
-                    brightSnr=brightSnr,
-                    medianRef=medianAstromscatterRef, matchRef=matchRef)
-    checkPhotometry(struct.snr, magavg, mmagrms, dist, match,
-                    brightSnr=brightSnr,
-                    medianRef=medianPhotoscatterRef, matchRef=matchRef)
+    astromScatter, astromFitParams = \
+        checkAstrometry(struct.snr, dist, match,
+                        brightSnr=brightSnr,
+                        medianRef=medianAstromscatterRef, matchRef=matchRef)
+    photScatter, photFitParams = \
+        checkPhotometry(struct.snr, magavg, mmagerr, mmagrms, dist, match,
+                        brightSnr=brightSnr,
+                        medianRef=medianPhotoscatterRef, matchRef=matchRef)
     if makePlot:
         plotAstrometry(dist, magavg, struct.snr,
+                       fit_params=astromFitParams,
                        brightSnr=brightSnr, outputPrefix=outputPrefix)
         plotPhotometry(magavg, struct.snr, mmagerr, mmagrms,
+                       fit_params=photFitParams,
                        brightSnr=brightSnr, filterName=filterName, outputPrefix=outputPrefix)
 
     magKey = allMatches.schema.find("base_PsfFlux_mag").key
